requirements.txt
telethon>=1.34
pytesseract>=0.3.10
opencv-python-headless>=4.8
httpx>=0.25
python-dotenv>=1.0
